Fixed CSS and merge conflicts. Everything works on this.
